minor-projects/ml_prediction_models/requirements.txt
numpy
pandas
scikit-learn
python-dotenv
flask